% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Bcells}
\alias{Bcells}
\title{flowSet of B cells.}
\format{A flowSet containing 3 flowFrames}
\usage{
data(Bcells)
}
\description{
This data set contain three flowFrame created by subsetting three fcs files
of and Aging study made in Singapore. The samples were stained with a panel
aimed to identify B cell subpopulations.
The data is stored as a flowSet object, a class implemented in the flowCore
package to handle fcs files in R.
}
\keyword{datasets}

