% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto-qc.R
\name{flow_auto_qc}
\alias{flow_auto_qc}
\title{Automatic quality control on .fcs files.}
\usage{
flow_auto_qc(fcsfiles, remove_from = "all", highQ_suffix = NULL,
  report = TRUE, report_suffix = NULL, save_highQ = TRUE,
  save_lowQ = TRUE, lowQ_suffix = NULL, timeCh = NULL,
  second_fractionFR = 0.1, long_periodFR = 1, periodFR = 30,
  alphaFR = 0.05, ChRemoveFS = NULL, outlierFS = FALSE,
  pen_valueFS = 200, max_cptFS = 3, ChFM = NULL, sideFM = "both")
}
\arguments{
\item{fcsfiles}{It can be a character vector with the filenames of the .fcs files,
a flowSet or a flowFrame.}

\item{remove_from}{Decide which steps have to be considered for the removal of
anomalies. The full quality control is performed by default with the option \code{"all"}.
Other options are: \code{'FR_FS' | 'FR_FM' | 'FS_FM' | 'FR' | 'FS' | 'FM'}.
The abbreviation \emph{FR} commands the QC for flow rate,
\emph{FS} commands the QC for signal acquisition, and \emph{FM} commands
the QC for the margins of the dynamic range.}

\item{highQ_suffix}{Character string to add to the original filenames
to name the new .fcs files containing the cells that passed the quality control.
The default is \code{NULL}, that automatically adds the suffix \emph{_highQ}
to the original name.}

\item{report}{logical indicating whether the report has to be released.
The default is \code{TRUE}.}

\item{report_suffix}{Character string to adds to the original filenames
to name the reports that will be released by the quality control.
The default is \code{NULL}, that automatically add the suffix \emph{_reportQC}
to the original name.}

\item{save_highQ}{logical indicating whether the high quality .fcs files have to be
saved. The default is \code{TRUE}.}

\item{save_lowQ}{logical indicating whether the low quality .fcs files have to be
saved. The default is \code{TRUE}.}

\item{lowQ_suffix}{Character string to add to the original filenames
to name the new .fcs files containing the cells that did not pass the quality control.
The default is \code{NULL}, that automatically adds the suffix \emph{_LowQ}
to the original name.}

\item{timeCh}{Character string corresponding to the name of the Time Channel
in the set of .fcs files. By default is \code{NULL} and the name is retrieved
automatically.}

\item{second_fractionFR}{The fraction of a second that is used to split
the time channel to recreate the flow rate. The fraction used by default is 1/10 of a second.}

\item{long_periodFR}{An integer indicating the number of sections the flow rate
should be split into and indipendently analysed. The default setting is \code{1}
and the flow rate is analyzed in its entirety.}

\item{periodFR}{An integer that indicates the number of segments a long period
in the flow rate has to be split into in order for the anomaly to be detected.
The default value is \code{30}.}

\item{alphaFR}{The level of statistical significance used to
accept anomalies. The default value is \code{0.05}.}

\item{ChRemoveFS}{Add a character vector with the names or
the pattern of the names of the channels that you do not want to include
in the signal quality control. The default option is \code{NULL}. If you want
to exclude the scatter parameters from the analysis use \code{c("FCS", "SSC")}.}

\item{outlierFS}{logical indicating whether to remove outliers from the intensity values
before performing the changepoint analysis. The default is \code{FALSE}.}

\item{pen_valueFS}{The value of the penalty for the changepoint detection
algorithm. This can be a numeric value or text giving the formula to use,
for instance it is possible to use the character string "1.5*log(n)",
where n indicates the number of cells in the .fcs file. The higher the
penalty value the less strict is the detection of the anomalies.
The default value is \code{200}.}

\item{max_cptFS}{The maximum number of changepoints that can be detected
for each channel. The default value is \code{3}.}

\item{ChFM}{A character vector that indicates which channels
need to undergo quality control of the dynamic range. The default
option is \code{NULL} and with it all the channels are selected for the analysis.}

\item{sideFM}{Select if the checking of the dynamic range should be performed in the
lower side, upper side or both sides of the acquitision range. Choose between:
\code{'both' | 'upper' | 'lower'}. The default value is \code{both}.}
}
\value{
It will return the .fcs files containing only the cells that
passed the quality checks. By default it will return a report indicating the cells
that were removed in the flow rate, signal acquisition over time and anomalous marginal
events on the dynamic range. It will also return a new .fcs file containing
only the cells that did not pass the quality check.
}
\description{
For a set of .fcs files, \emph{flow_auto_qc} performs the quality control in three steps.
It automatically detects and removes anomalies in the flow rate, in the
signal acquisition and in the dynamic range.
}
\examples{

## a sample file
data(Bcells)

## quality control
flow_auto_qc(Bcells[[1]], report = FALSE,  save_highQ = FALSE, save_lowQ = FALSE)

}
\author{
Gianni Monaco with substantial contribution from Chen Hao
}

