% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{flow_iQC}
\alias{flow_iQC}
\title{A Shiny app to perform interactive quality control on Flow Cytometry Data}
\usage{
flow_iQC()
}
\description{
This function allows to interactively perform the quality control of .fcs files in a Shiny
app. The quality control framework of the Shiny app is equivalent to the
one of the \code{flow_auto_qc} function. It is composed of three main steps that allow
to check anomalies in the 1) flow rate, 2) signal acquisition, 3) dynamic range.
}
\examples{
if (interactive()) flowAI::flow_iQC()
}
\author{
Chen Hao
}

